Usmap rewrite + versioning
