+++ CUE4Parse/MappingsProvider/Usmap/EUsmapCompressionMethod.cs
@@ -1,0 +1,10 @@
+﻿namespace CUE4Parse.MappingsProvider.Usmap;
+
+public enum EUsmapCompressionMethod : byte
+{
+    None,
+    Oodle,
+    Brotli,
+
+    Unknown = 0xFF
+}
